==> 4a_Production_of_GUI_in1.ipynb <==
import os

import datetime

import numpy as np
import pandas as pd

from pandas import Series
from pandas import DataFrame

from scipy.stats import norm

import matplotlib.pyplot as plt

downweight_granularity=['dtime','Direction','dday','month']

Data=pd.read_csv('R_Output_Test_Pax.csv',sep=',')
Data=Data[['DepDate','yday','dtime','Direction','month','dday','Dprio','NumPax']]

# Data Import

Data=pd.read_csv('R_Output_Test_Pax.csv',sep=',')
Data=Data[['DepDate','yday','dtime','Direction','month','dday','Dprio','NumPax']]

Map_Dates=DataFrame([Data['DepDate'].unique().tolist(),[pd.to_datetime(x) for x in Data['DepDate'].unique().tolist()]]).transpose()
Map_Dates.columns=['DepDate','DepDate_dt']
Map_Dates['DepDate']=Map_Dates['DepDate'].astype('str')

Data=Data.merge(Map_Dates,on=['DepDate'])
Data['DepDate']=Data['DepDate_dt']
Data=Data[[x for x in Data.columns if x!='DepDate_dt']]

Data['TicketDate']=Data['DepDate']-Data['Dprio'].apply(lambda x: datetime.timedelta(x))

Data=Data.loc[Data['TicketDate']<=pd.to_datetime(datetime.datetime.today().strftime('%Y-%m-%d')),:]

Data=Data[[x for x in Data.columns if x!='TicketDate']]

Data['Dprio']=-Data['Dprio']
Data=Data.set_index(['DepDate','yday','dtime','Direction','month','dday','Dprio'])
Data=Data.groupby(level=[0,1,2,3,4,5,6]).sum().groupby(level=[0,1,2,3,4,5])['NumPax'].cumsum().reset_index()

Data['Dprio']=-Data['Dprio']

IdealCurve=pd.read_csv('IdealCurve.csv')

# deptime=[str(x/60).split('.')[0]+':'+str(round(float('0.'+str(x/60).split('.')[1])*60)) for x in IdealCurve['dtime'].unique()]
# deptime=[x+'0' if len(x.split(':')[1])==1 else x for x in deptime]
# deptime=['0'+x if len(x.split(':')[0])==1 else x for x in deptime]

# Map_DepTime=DataFrame([IdealCurve['dtime'].unique(),[x+'0' if len(x.split(':')[1])==1 else x for x in deptime]]).transpose()
# Map_DepTime.columns=['dtime','deptime']
# Map_DepTime['dtime']=Map_DepTime['dtime'].astype('int')

# IdealCurve=IdealCurve.merge(Map_DepTime,on='dtime')

Dropdown_comb=Data[['DepDate']+downweight_granularity].drop_duplicates()
Dropdown_comb.index=range(Dropdown_comb.shape[0])
Dprio_all=Data['Dprio'].unique().tolist()

Dropdown_comb=Dropdown_comb.loc[np.repeat(Dropdown_comb.index,len(Dprio_all)),:]
Dropdown_comb.index=range(Dropdown_comb.shape[0])

Dropdown_comb=pd.concat([Dropdown_comb,Series(Dprio_all*Dropdown_comb.drop_duplicates().shape[0])],axis=1)
Dropdown_comb=Dropdown_comb.rename(columns={0: 'Dprio'})

Dropdown_comb=Dropdown_comb.merge(IdealCurve,on=downweight_granularity+['Dprio'])

Dropdown_comb.to_csv('IdealCurve_forPlots.csv',index=False)

#Dropdown_comb.loc[Dropdown_comb['DepDate']=='2019-06-29',:]

Data=Data[['DepDate','dtime','Direction','dday','Dprio','NumPax','Ideal_lower','Ideal_average_downweighted','Ideal_average_full','Ideal_upper']]\
.sort_values(by=['DepDate','Direction','dtime','Dprio'])

Data['NumPax']=Data['NumPax'].apply(lambda x: int(round(x)))
Data['Ideal_lower']=Data['Ideal_lower'].apply(lambda x: int(round(x)))
Data['Ideal_average_downweighted']=Data['Ideal_average_downweighted'].apply(lambda x: int(round(x)))
Data['Ideal_average_full']=Data['Ideal_average_full'].apply(lambda x: int(round(x)))
Data['Ideal_upper']=Data['Ideal_upper'].apply(lambda x: int(round(x)))

deptime=[str(x/60).split('.')[0]+':'+str(round(float('0.'+str(x/60).split('.')[1])*60)) for x in Data['dtime'].unique()]
deptime=[x+'0' if len(x.split(':')[1])==1 else x for x in deptime]
deptime=['0'+x if len(x.split(':')[0])==1 else x for x in deptime]

Map_DepTime=DataFrame([Data['dtime'].unique(),[x+'0' if len(x.split(':')[1])==1 else x for x in deptime]]).transpose()
Map_DepTime.columns=['dtime','deptime']
Map_DepTime['dtime']=Map_DepTime['dtime'].astype('int')

Data=Data.merge(Map_DepTime,on='dtime')

Data['dtime']=Data['deptime']
Data=Data[[x for x in Data.columns if x!='deptime']]

Data=Data.sort_values(by=['DepDate','dtime'])
Data=Data.rename(columns={'NumPax': 'Actual Bookings', 'Ideal_lower': 'Ramp-up frontier',
'Ideal_average_downweighted': 'Ideal curve (80% LF)', 'Ideal_average_full': 'Ideal curve (100% LF)',
'Ideal_upper': 'Phase-down frontier'})

GroupPax=pd.read_csv('GroupPax.csv').rename(columns={'Group pax': 'Group Pax'})

GroupPax['DepDate']=pd.to_datetime(GroupPax['DepDate'])
GroupPax['dtime']=GroupPax['dtime'].astype('str')
GroupPax['Dprio']=GroupPax['Dprio'].astype('float')

Data=Data.merge(GroupPax[['DepDate','dtime','Dprio','Group Pax']],on=['DepDate','dtime','Dprio'])

Capacity=pd.read_csv('Capacity.csv')

Capacity['DepDate']=pd.to_datetime(Capacity['DepDate'])
Capacity['dtime']=Capacity['dtime'].astype('str')
Capacity['Dprio']=Capacity['Dprio'].astype('float')

Data=Data.merge(Capacity[['DepDate','dtime','Dprio','Cabin Capacity']],on=['DepDate','dtime','Dprio'])

Data=Data.to_csv('GUI_in.csv',index=False)
